requirements.txt
asgiref==3.8.1
build==1.2.2.post1
Django==5.2
django-model-utils==5.0.0
django-polymorphic==3.1.0
packaging==25.0
pyproject_hooks==1.2.0
sqlparse==0.5.3
